requirements.txt
flask>=2.0.0,<3.0.0
requests>=2.25.0,<3.0.0
langchain>=0.0.100,<0.1.0
openai>=0.27.0,<0.28.0
faiss-cpu>=1.7.0,<1.8.0
python-dotenv>=0.19.0,<0.20.0
flask-cors>=3.0.0,<4.0.0
tqdm>=4.0.0,<5.0.0
